refact: split Flex and Grid to server and client and detect automatically. feat: custom cursor.
