+++ packages/core/src/components/Flex.tsx
@@ -1,429 +1,28 @@
-"use client";
+import React, { forwardRef } from "react";
+import { ClientFlex, ServerFlex } from ".";
+import { FlexProps, StyleProps, SpacingProps, SizeProps, CommonProps, DisplayProps } from "../interfaces";
 
-import classNames from "classnames";
-import { CSSProperties, forwardRef, useEffect, useRef, useCallback } from "react";
+interface SmartFlexProps extends FlexProps, StyleProps, SpacingProps, SizeProps, CommonProps, DisplayProps {}
 
-import {
-  CommonProps,
-  DisplayProps,
-  FlexProps,
-  SizeProps,
-  SpacingProps,
-  StyleProps,
-} from "../interfaces";
-import { ColorScheme, ColorWeight, SpacingToken, TextVariant } from "../types";
-
-interface ComponentProps
-  extends FlexProps,
-    SpacingProps,
-    SizeProps,
-    StyleProps,
-    CommonProps,
-    DisplayProps {}
-
-const Flex = forwardRef<HTMLDivElement, ComponentProps>(
-  (
-    {
-      as: Component = "div",
-      inline,
-      hide,
-      show,
-      dark,
-      light,
-      direction,
-      l,
-      m,
-      s,
-      wrap = false,
-      horizontal,
-      vertical,
-      flex,
-      textVariant,
-      textSize,
-      textWeight,
-      textType,
-      onBackground,
-      onSolid,
-      align,
-      top,
-      right,
-      bottom,
-      left,
-      padding,
-      paddingLeft,
-      paddingRight,
-      paddingTop,
-      paddingBottom,
-      paddingX,
-      paddingY,
-      margin,
-      marginLeft,
-      marginRight,
-      marginTop,
-      marginBottom,
-      marginX,
-      marginY,
-      gap,
-      position = "relative",
-      center,
-      width,
-      height,
-      maxWidth,
-      minWidth,
-      minHeight,
-      maxHeight,
-      fit = false,
-      fitWidth = false,
-      fitHeight = false,
-      fill = false,
-      fillWidth = false,
-      fillHeight = false,
-      aspectRatio,
-      transition,
-      background,
-      solid,
-      opacity,
-      pointerEvents,
-      border,
-      borderTop,
-      borderRight,
-      borderBottom,
-      borderLeft,
-      borderStyle,
-      borderWidth,
-      radius,
-      topRadius,
-      rightRadius,
-      bottomRadius,
-      leftRadius,
-      topLeftRadius,
-      topRightRadius,
-      bottomLeftRadius,
-      bottomRightRadius,
-      overflow,
-      overflowX,
-      overflowY,
-      zIndex,
-      shadow,
-      cursor,
-      className,
-      style,
-      children,
-      ...rest
-    },
-    ref,
-  ) => {
-    const elementRef = useRef<HTMLDivElement>(null);
-    const appliedResponsiveStyles = useRef<Set<string>>(new Set());
-    const baseStyleRef = useRef<CSSProperties>({});
-    
-    const combinedRef = (node: HTMLDivElement) => {
-      elementRef.current = node;
-      if (typeof ref === 'function') {
-        ref(node);
-      } else if (ref) {
-        ref.current = node;
-      }
-    };
-    if (onBackground && onSolid) {
-      console.warn(
-        "You cannot use both 'onBackground' and 'onSolid' props simultaneously. Only one will be applied.",
-      );
-    }
-
-    if (background && solid) {
-      console.warn(
-        "You cannot use both 'background' and 'solid' props simultaneously. Only one will be applied.",
-      );
-    }
-
-    const getVariantClasses = (variant: TextVariant) => {
-      const [fontType, weight, size] = variant.split("-");
-      return [`font-${fontType}`, `font-${weight}`, `font-${size}`];
-    };
-
-    const sizeClass = textSize ? `font-${textSize}` : "";
-    const weightClass = textWeight ? `font-${textWeight}` : "";
-
-    const variantClasses = textVariant ? getVariantClasses(textVariant) : [sizeClass, weightClass];
-
-    let colorClass = "";
-    if (onBackground) {
-      const [scheme, weight] = onBackground.split("-") as [ColorScheme, ColorWeight];
-      colorClass = `${scheme}-on-background-${weight}`;
-    } else if (onSolid) {
-      const [scheme, weight] = onSolid.split("-") as [ColorScheme, ColorWeight];
-      colorClass = `${scheme}-on-solid-${weight}`;
-    }
-
-    const generateDynamicClass = (type: string, value: string | undefined) => {
-      if (!value) return undefined;
-
-      if (value === "transparent") {
-        return `transparent-border`;
-      }
-
-      if (["surface", "page", "overlay"].includes(value)) {
-        return `${value}-${type}`;
-      }
-
-      const parts = value.split("-");
-      if (parts.includes("alpha")) {
-        const [scheme, , weight] = parts;
-        return `${scheme}-${type}-alpha-${weight}`;
-      }
-
-      const [scheme, weight] = value.split("-") as [ColorScheme, ColorWeight];
-      return `${scheme}-${type}-${weight}`;
-    };
-
-    const classes = classNames(
-      inline ? "display-inline-flex" : "display-flex",
-      position && `position-${position}`,
-      l?.position && `l-position-${l.position}`,
-      m?.position && `m-position-${m.position}`,
-      s?.position && `s-position-${s.position}`,
-      padding && `p-${padding}`,
-      paddingLeft && `pl-${paddingLeft}`,
-      paddingRight && `pr-${paddingRight}`,
-      paddingTop && `pt-${paddingTop}`,
-      paddingBottom && `pb-${paddingBottom}`,
-      paddingX && `px-${paddingX}`,
-      paddingY && `py-${paddingY}`,
-      margin && `m-${margin}`,
-      marginLeft && `ml-${marginLeft}`,
-      marginRight && `mr-${marginRight}`,
-      marginTop && `mt-${marginTop}`,
-      marginBottom && `mb-${marginBottom}`,
-      marginX && `mx-${marginX}`,
-      marginY && `my-${marginY}`,
-      gap === "-1"
-        ? direction === "column" || direction === "column-reverse"
-          ? "g-vertical--1"
-          : "g-horizontal--1"
-        : gap && `g-${gap}`,
-      top ? `top-${top}` : (position === "sticky" ? "top-0" : undefined),
-      right && `right-${right}`,
-      bottom && `bottom-${bottom}`,
-      left && `left-${left}`,
-      generateDynamicClass("background", background),
-      generateDynamicClass("solid", solid),
-      generateDynamicClass(
-        "border",
-        border || borderTop || borderRight || borderBottom || borderLeft,
-      ),
-      (border || borderTop || borderRight || borderBottom || borderLeft) &&
-        !borderStyle &&
-        "border-solid",
-      border && !borderWidth && "border-1",
-      (borderTop || borderRight || borderBottom || borderLeft) && "border-reset",
-      borderTop && "border-top-1",
-      borderRight && "border-right-1",
-      borderBottom && "border-bottom-1",
-      borderLeft && "border-left-1",
-      borderWidth && `border-${borderWidth}`,
-      borderStyle && `border-${borderStyle}`,
-      radius === "full" ? "radius-full" : radius && `radius-${radius}`,
-      topRadius && `radius-${topRadius}-top`,
-      rightRadius && `radius-${rightRadius}-right`,
-      bottomRadius && `radius-${bottomRadius}-bottom`,
-      leftRadius && `radius-${leftRadius}-left`,
-      topLeftRadius && `radius-${topLeftRadius}-top-left`,
-      topRightRadius && `radius-${topRightRadius}-top-right`,
-      bottomLeftRadius && `radius-${bottomLeftRadius}-bottom-left`,
-      bottomRightRadius && `radius-${bottomRightRadius}-bottom-right`,
-      direction && `flex-${direction}`,
-      l?.direction && `l-flex-${l.direction}`,
-      m?.direction && `m-flex-${m.direction}`,
-      s?.direction && `s-flex-${s.direction}`,
-      pointerEvents && `pointer-events-${pointerEvents}`,
-      transition && `transition-${transition}`,
-      hide && "flex-hide",
-      show && "flex-show",
-      l?.hide && "l-flex-hide",
-      l?.show && "l-flex-show",
-      m?.hide && "m-flex-hide",
-      m?.show && "m-flex-show",
-      s?.hide && "s-flex-hide",
-      s?.show && "s-flex-show",
-      opacity && `opacity-${opacity}`,
-      wrap && "flex-wrap",
-      overflow && `overflow-${overflow}`,
-      overflowX && `overflow-x-${overflowX}`,
-      overflowY && `overflow-y-${overflowY}`,
-      flex && `flex-${flex}`,
-      horizontal &&
-        (direction === "row" || direction === "row-reverse" || direction === undefined
-          ? `justify-${horizontal}`
-          : `align-${horizontal}`),
-      vertical &&
-        (direction === "row" || direction === "row-reverse" || direction === undefined
-          ? `align-${vertical}`
-          : `justify-${vertical}`),
-      l?.horizontal &&
-        (l?.direction === "row" || l?.direction === "row-reverse" || l?.direction === undefined
-          ? `l-justify-${l.horizontal}`
-          : `l-align-${l.horizontal}`),
-      l?.vertical &&
-        (l?.direction === "row" || l?.direction === "row-reverse" || l?.direction === undefined
-          ? `l-align-${l.vertical}`
-          : `l-justify-${l.vertical}`),
-      m?.horizontal &&
-        (m?.direction === "row" || m?.direction === "row-reverse" || m?.direction === undefined
-          ? `m-justify-${m.horizontal}`
-          : `m-align-${m.horizontal}`),
-      m?.vertical &&
-        (m?.direction === "row" || m?.direction === "row-reverse" || m?.direction === undefined
-          ? `m-align-${m.vertical}`
-          : `m-justify-${m.vertical}`),
-      s?.horizontal &&
-        (s?.direction === "row" || s?.direction === "row-reverse" || s?.direction === undefined
-          ? `s-justify-${s.horizontal}`
-          : `s-align-${s.horizontal}`),
-      s?.vertical &&
-        (s?.direction === "row" || s?.direction === "row-reverse" || s?.direction === undefined
-          ? `s-align-${s.vertical}`
-          : `s-justify-${s.vertical}`),
-      center && "center",
-      fit && "fit",
-      fitWidth && "fit-width",
-      fitHeight && "fit-height",
-      fill && "fill",
-      fillWidth && !minWidth && "min-width-0",
-      fillHeight && !minHeight && "min-height-0",
-      fill && "min-height-0",
-      fill && "min-width-0",
-      (fillWidth || maxWidth) && "fill-width",
-      (fillHeight || maxHeight) && "fill-height",
-      shadow && `shadow-${shadow}`,
-      zIndex && `z-index-${zIndex}`,
-      textType && `font-${textType}`,
-      cursor && `cursor-${cursor}`,
-      dark && "dark-flex",
-      light && "light-flex",
-      colorClass,
-      className,
-      ...variantClasses,
-    );
-
-    const parseDimension = (
-      value: number | SpacingToken | undefined,
-      type: "width" | "height",
-    ): string | undefined => {
-      if (value === undefined) return undefined;
-      if (typeof value === "number") return `${value}rem`;
-      if (
-        [
-          "0",
-          "1",
-          "2",
-          "4",
-          "8",
-          "12",
-          "16",
-          "20",
-          "24",
-          "32",
-          "40",
-          "48",
-          "56",
-          "64",
-          "80",
-          "104",
-          "128",
-          "160",
-        ].includes(value)
-      ) {
-        return `var(--static-space-${value})`;
-      }
-      if (["xs", "s", "m", "l", "xl"].includes(value)) {
-        return `var(--responsive-${type}-${value})`;
-      }
-      return undefined;
-    };
-
-    const applyResponsiveStyles = useCallback(() => {
-      if (!elementRef.current) return;
-      
-      const element = elementRef.current;
-      const width = window.innerWidth;
-      
-      // Update base styles when style prop changes
-      if (style) {
-        baseStyleRef.current = { ...style };
-      }
-      
-      // Determine which responsive props to apply based on breakpoint
-      let currentResponsiveProps: any = null;
-      if (width >= 1025 && l) {
-        currentResponsiveProps = l;
-      } else if (width >= 769 && width <= 1024 && m) {
-        currentResponsiveProps = m;
-      } else if (width <= 768 && s) {
-        currentResponsiveProps = s;
-      }
-      
-      // Clear only responsive styles, not base styles
-      appliedResponsiveStyles.current.forEach(key => {
-        (element.style as any)[key] = '';
-      });
-      appliedResponsiveStyles.current.clear();
-      
-      // Reapply base styles
-      if (baseStyleRef.current) {
-        Object.entries(baseStyleRef.current).forEach(([key, value]) => {
-          (element.style as any)[key] = value;
-        });
-      }
-      
-      // Apply new responsive styles if we have them for current breakpoint
-      if (currentResponsiveProps) {
-        if (currentResponsiveProps.style) {
-          Object.entries(currentResponsiveProps.style).forEach(([key, value]) => {
-            (element.style as any)[key] = value;
-            appliedResponsiveStyles.current.add(key);
-          });
-        }
-        if (currentResponsiveProps.aspectRatio) {
-          element.style.aspectRatio = currentResponsiveProps.aspectRatio;
-          appliedResponsiveStyles.current.add('aspect-ratio');
-        }
-      }
-    }, [l, m, s, style]);
-
-    useEffect(() => {
-      applyResponsiveStyles();
-      
-      const handleResize = () => {
-        applyResponsiveStyles();
-      };
-      
-      window.addEventListener('resize', handleResize);
-      
-      return () => {
-        window.removeEventListener('resize', handleResize);
-      };
-    }, [applyResponsiveStyles]);
-
-    const combinedStyle: CSSProperties = {
-      maxWidth: parseDimension(maxWidth, "width"),
-      minWidth: parseDimension(minWidth, "width"),
-      minHeight: parseDimension(minHeight, "height"),
-      maxHeight: parseDimension(maxHeight, "height"),
-      width: parseDimension(width, "width"),
-      height: parseDimension(height, "height"),
-      aspectRatio: aspectRatio,
-      textAlign: align,
-      ...style,
-    };
-
-    return (
-      <Component ref={combinedRef} className={classes} style={combinedStyle} {...rest}>
-        {children}
-      </Component>
-    );
-  },
-);
+const Flex = forwardRef<HTMLDivElement, SmartFlexProps>(({ cursor, l, m, s, style, ...props }, ref) => {
+  // Use client component if custom cursor (ReactNode) is provided
+  if (typeof cursor === 'object' && cursor) {
+    return <ClientFlex ref={ref} cursor={cursor} l={l} m={m} s={s} style={style} {...props} />;
+  }
+  
+  // Use client component if responsive props are provided
+  if (l || m || s) {
+    return <ClientFlex ref={ref} cursor={cursor} l={l} m={m} s={s} style={style} {...props} />;
+  }
+  
+  // Use client component if dynamic style is provided (requires client-side updates)
+  if (style && typeof style === 'object' && Object.keys(style).length > 0) {
+    return <ClientFlex ref={ref} cursor={cursor} l={l} m={m} s={s} style={style} {...props} />;
+  }
+  
+  // Use server component for string cursors or no cursor, no responsive props, no dynamic styles
+  return <ServerFlex ref={ref} cursor={cursor} {...props} />;
+});
 
 Flex.displayName = "Flex";
 export { Flex };
