+++ packages/core/src/components/index.ts
@@ -13,6 +13,7 @@ export * from "./Button";
 export * from "./Carousel";
 export * from "./Card";
 export * from "./CursorCard";
+export * from "./Cursor";
 export * from "./Column";
 export * from "./Checkbox";
 export * from "./Chip";
@@ -31,10 +32,14 @@ export * from "./EmojiPicker";
 export * from "./EmojiPickerDropdown";
 export * from "./Fade";
 export * from "./Feedback";
+export * from "./ServerFlex";
+export * from "./ClientFlex";
 export * from "./Flex";
 export * from "./FlipFx";
 export * from "./FocusTrap";
 export * from "./GlitchFx";
+export * from "./ServerGrid";
+export * from "./ClientGrid";
 export * from "./Grid";
 export * from "./Heading";
 export * from "./HoloFx";
